feat: enrich tools dashboard with live data, API proxy, and bug fixes
- Expand expression tester to 18 FEL examples across all categories
- Build export cards dynamically from mapping docs with rule visualization
- Generate rich changelog diffs covering items/binds/shapes/optionSets/screener
- Add registry card detail sections (baseType, constraints, params, deprecation)
- Fix d3 zoom to use a single transform group instead of all g elements
- Add Vite proxy (/api → localhost:8000) so dashboard works through any host
- Fix Rating component to respect integer dataType (round half values)
- Fix JsonAdapter to handle Decimal serialization from mapping engine
- Update instance data source to use local API server with inline fallback
- Update README with server API docs, tools dashboard guide, and setup changes

+++ tests/e2e/playwright/components/interactive-components.spec.ts
@@ -287,7 +287,7 @@ test.describe('Components: Progressive Component Rendering', () => {
             el.definition = {
                 "$formspec": "1.0", "url": "http://example.org/test",
                 "version": "1.0.0", "title": "Rating Half Test",
-                "items": [{ "key": "score", "type": "field", "dataType": "integer", "label": "Score" }]
+                "items": [{ "key": "score", "type": "field", "dataType": "decimal", "label": "Score" }]
             };
             el.componentDocument = {
                 "$formspecComponent": "1.0",
